Correct examples to follow more restrictive 'no multiply defined path' model.

+++ architecture/faust/gui/Soundfile.h
@@ -190,7 +190,7 @@ class SoundfileReader {
             int total_length = 0;
             
             // Compute total length and channels max of all files
-            for (size_t i = 0; i < path_name_list.size(); i++) {
+            for (int i = 0; i < int(path_name_list.size()); i++) {
                 int chan, length;
                 if (path_name_list[i] == "__empty_sound__") {
                     length = BUFFER_SIZE;
@@ -212,7 +212,7 @@ class SoundfileReader {
             int offset = 0;
             
             // Read all files
-            for (size_t i = 0; i < path_name_list.size(); i++) {
+            for (int i = 0; i < int(path_name_list.size()); i++) {
                 if (path_name_list[i] == "__empty_sound__") {
                     emptyFile(soundfile, i, offset);
                 } else {
@@ -221,7 +221,7 @@ class SoundfileReader {
             }
             
             // Complete with empty parts
-            for (int i = path_name_list.size(); i < MAX_SOUNDFILE_PARTS; i++) {
+            for (int i = int(path_name_list.size()); i < MAX_SOUNDFILE_PARTS; i++) {
                 emptyFile(soundfile, i, offset);
             }
